Move watchlist actions to nested controller and use ES revenue
Move add/update/remove watchlist actions out of Admin::UsersController
into a nested singular resource at admin/users/:user_external_id/watchlist
backed by Admin::Users::WatchlistsController. Compute revenue_cents via
the same Elasticsearch-backed sales_cents_total used at the top of the
admin user page so the snapshot matches what reviewers already see, and
drop the unused revenue_progress_percent helper.
